% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ANALYTES_EMM}
\alias{ANALYTES_EMM}
\title{Plasma clinical analytes estimated marginal means}
\format{
An object of class \code{list} of length 9.
}
\usage{
ANALYTES_EMM
}
\description{
A list of \code{emmGrid} objects, one for each clinical analyte.
}
\examples{
names(ANALYTES_EMM) # clinical analytes

str(ANALYTES_EMM)

# Print one of the emmGrid objects
summary(ANALYTES_EMM[["Corticosterone"]])

}
\seealso{
\code{\link{ANALYTES}}, \code{\link{ANALYTES_STATS}},
\code{\link[emmeans]{emmeans}}
}
\keyword{datasets}
