% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{BASELINE_EMM}
\alias{BASELINE_EMM}
\title{Baseline (pre-training) measures estimated marginal means}
\format{
An object of class \code{list} of length 7.
}
\usage{
BASELINE_EMM
}
\description{
A list of \code{emmGrid} objects: NMR mass, lean mass, fat mass,
\% lean mass, \% fat mass; absolute VO\eqn{_2}max; and relative
VO\eqn{_2}max.
}
\examples{
names(BASELINE_EMM) # available measures

str(BASELINE_EMM)

# Print one of the emmGrid objects
summary(BASELINE_EMM[["NMR Body Mass"]])

}
\seealso{
\code{\link{BASELINE_STATS}}, \code{\link[emmeans]{emmeans}}
}
\keyword{datasets}
