% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{BODY_MASSES}
\alias{BODY_MASSES}
\title{Body mass recorded on different days}
\format{
An object of class \code{data.frame} with 282 rows and 12 columns:

\describe{
\item{pid}{character; randomly generated 8-digit identifier used in
linkage to phenotypic data. All samples from the same animal have the same
PID.}
\item{iowa_id}{character; unique ID used for each rat at the Iowa PASS.}
\item{age}{factor; the age of the rat at the beginning of the training
protocol. Levels "6M" (Adult) and "18M" (Aged).}
\item{sex}{factor; the sex of the rat with levels "Female" and "Male".}
\item{group}{factor; exercise training group. Either "SED" (sedentary) or
the number of weeks of training ("1W", "2W", "4W", "8W").}
\item{arrival_body_mass}{numeric; body mass (g) recorded upon arrival to
the facility.}
\item{famil_body_mass}{numeric; body mass (g) following the familiarization
period.}
\item{nmr_pre_body_mass}{numeric; body mass (g) recorded on the same day as
NMR body composition measures (pre-training).}
\item{nmr_post_body_mass}{numeric; body mass (g) recorded on the same day
as NMR body composition measures (post-training).}
\item{vo2_pre_body_mass}{numeric; body mass (g) recorded on the same day as
the VO\eqn{_2}max test (pre-training).}
\item{vo2_post_body_mass}{numeric; body mass (g) recorded on the same day
as the VO\eqn{_2}max test (post-training).}
\item{term_body_mass}{numeric; body mass (g) recorded on the day of
sacrifice.}
}
}
\usage{
BODY_MASSES
}
\description{
Measures of total body mass (in grams) recorded on different
days (pre- and post-training measures included, if applicable): NMR day,
VO\eqn{_2}max testing day, and day of sacrifice (terminal).
}
\examples{
head(BODY_MASSES)

}
\keyword{datasets}
