% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{DISSECTION_TIMES}
\alias{DISSECTION_TIMES}
\title{Tissue dissection times}
\format{
An object of class \code{data.frame} with 5450 rows and 11 columns:

\describe{
\item{pid}{character; randomly generated 8-digit identifier used in
linkage to phenotypic data. All samples from the same animal have the same
PID.}
\item{age}{factor; the age of the rat at the beginning of the training
protocol. Levels: "6M" (Adult) and "18M" (Aged).}
\item{sex}{factor; the sex of the rat with levels "Female" and "Male".}
\item{group}{factor; exercise training group. Either "SED" (sedentary) or
the number of weeks of training ("1W", "2W", "4W", "8W").}
\item{sampletypedescription}{character; sample type (usually a tissue)
descriptions.}
\item{aliquotdescription}{factor; sample type descriptions for plotting.}
\item{t_collection}{Period; time of tissue collection relative to midnight
("0S").}
\item{t_death}{Period; time of death via heart puncture relative to
midnight ("0S").}
\item{t_freeze}{Period; time of tissue freezing relative to midnight
("0S").}
\item{collect_before_death}{logical; whether the tissue was collected
before the rat was sacrificed.}
\item{t_diff}{integer; seconds elapsed between time of freezing and either
time of collection (if \code{collect_before_death}) or time of death
(if \code{!collect_before_death}).}
}
}
\usage{
DISSECTION_TIMES
}
\description{
Tissue dissection times.
}
\keyword{datasets}
