% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{FIBER_AREA_EMM}
\alias{FIBER_AREA_EMM}
\title{Muscle fiber cross-sectional area estimated marginal means}
\format{
An object of class \code{emmGrid} of length 1.
}
\usage{
FIBER_AREA_EMM
}
\description{
An \code{emmGrid} object containing estimated marginal means for
the muscle fiber cross-sectional area data by fiber type.
}
\examples{
str(FIBER_AREA_EMM)

}
\seealso{
\code{\link{FIBER_TYPES}}, \code{\link{FIBER_AREA_STATS}},
\code{\link[emmeans]{emmeans}}
}
\keyword{datasets}
