% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{FIBER_AREA_STATS}
\alias{FIBER_AREA_STATS}
\title{Statistical analysis results of fiber cross-sectional area}
\format{
An object of class \code{data.frame} with 56 rows and 21 columns:

\describe{
\item{response}{character; the measurement being tested. All "Mean Fiber
Area".}
\item{age}{factor; the age of the rat at the beginning of the training
protocol. Levels: "6M" (Adult) and "18M" (Aged).}
\item{sex}{factor; the sex of the rat with levels "Female" and "Male".}
\item{muscle}{factor; the muscle that was sampled with levels "LG" (lateral
gastrocnemius), "MG" (medial gastrocnemius), "PL" (plantaris), and "SOL"
(soleus).}
\item{type}{factor; the muscle fiber type with levels "I", "IIa", "IIx",
and "IIb".}
\item{contrast}{factor; the comparison between groups. All contrasts are
ratios between the 8W and SED group means ("8W / SED").}
\item{estimate_type}{character; interpretation of the value in the
\code{estimate} column. All "ratio" (ratio between group means, as
specified by \code{contrast}).}
\item{null}{numeric; the value of the estimate under the null hypothesis.}
\item{estimate}{numeric; ratio between the means of the groups as specified
by \code{contrast}.}
\item{SE}{numeric; the standard error of the estimate.}
\item{lower.CL}{numeric; lower bound of the 95\% confidence interval.}
\item{upper.CL}{numeric; upper bound of the 95\% confidence interval.}
\item{statistic_type}{character; the type of statistical test. All "t"
(Student's t-statistic).}
\item{statistic}{numeric; the value of the test statistic.}
\item{df}{numeric; the number of residual degrees of freedom.}
\item{p.value}{numeric; the p-value associated with the statistical
test.}
\item{p.adj}{numeric; the Holm-adjusted p-value.}
\item{signif}{logical; \code{TRUE} if \code{p.adj} < 0.05.}
\item{model_type}{character; the statistical model used for testing. All
"lme" (linear mixed effects model).}
\item{fixed}{character; the fixed-effects component of the model formula.
Describes the response variable (and any transformations) and all predictor
variables that were included. Equivalent to \code{formula}.}
\item{random}{character; the random component of the model formula.
Specifies the random effects for mixed effects models. All "~1 | pid"
(random intercept for each rat).}
}
}
\usage{
FIBER_AREA_STATS
}
\description{
Statistical analysis results of the fiber cross-sectional area
data by muscle and fiber type.
}
\examples{
head(FIBER_AREA_STATS)

}
\seealso{
\code{\link{FIBER_TYPES}}, \code{\link{FIBER_AREA_EMM}}
}
\keyword{datasets}
