% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{FIBER_COUNT_EMM}
\alias{FIBER_COUNT_EMM}
\title{Muscle fiber count estimated marginal means}
\format{
An object of class \code{list} of length 4.
}
\usage{
FIBER_COUNT_EMM
}
\description{
An \code{emmGrid} object containing estimated marginal means for
the muscle fiber count data. Note that the counts have been transformed to
isometric log-ratio (ilr) coordinates. See the vignette for details.
}
\keyword{datasets}
