% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{FIBER_COUNT_STATS}
\alias{FIBER_COUNT_STATS}
\title{Statistical analysis results of muscle fiber counts}
\format{
An object of class \code{data.frame} with 40 rows and 19 columns:

\describe{
\item{response}{character; the measurement being tested. All "Fiber
Count".}
\item{age}{factor; the age of the rat at the beginning of the training
protocol. Levels: "6M" (Adult) and "18M" (Aged).}
\item{sex}{factor; the sex of the rat with levels "Female" and "Male".}
\item{muscle}{factor; the muscle that was sampled with levels "LG" (lateral
gastrocnemius), "MG" (medial gastrocnemius), "PL" (plantaris), and "SOL"
(soleus).}
\item{balance}{numeric; Exploring Compositional Data with the
CoDa-Dendrogram: "the natural logarithm of the ratio of geometric means of
the parts in each group, normalised by a coefficient to guarantee unit
length of the vectors of the basis" (Pawlowsky-Glahn & Egozcue 2011).}
\item{contrast}{factor; the comparison between groups. All contrasts are
differences between the 8W and SED group means ("8W - SED").}
\item{estimate_type}{character; interpretation of the value in the
\code{estimate} column. All "difference of means" (difference between group
means, as specified by \code{contrast}).}
\item{estimate}{numeric; difference between the means of the groups as
specified by \code{contrast}.}
\item{SE}{numeric; the standard error of the estimate.}
\item{lower.CL}{numeric; lower bound of the 95\% confidence interval.}
\item{upper.CL}{numeric; upper bound of the 95\% confidence interval.}
\item{statistic_type}{character; the type of statistical test.}
\item{statistic}{numeric; the value of the test statistic.}
\item{df}{numeric; the number of residual degrees of freedom.}
\item{p.value}{numeric; the p-value associated with the statistical
test.}
\item{p.adj}{numeric; the adjusted p-value.}
\item{signif}{logical; \code{TRUE} if \code{p.adj} < 0.05.}
\item{model_type}{character; the statistical model used for testing.}
\item{formula}{character; the model formula. Includes the response
variable, any transformations, and predictors.}
}
}
\usage{
FIBER_COUNT_STATS
}
\description{
Statistical analysis results of muscle fiber counts.
}
\references{
Pawlowsky-Glahn, V., & Egozcue, J. J. (2011). Exploring
Compositional Data with the CoDa-Dendrogram. \emph{Austrian Journal of
Statistics, 40}(1&2), 103--113.
}
\keyword{datasets}
