% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{FIBER_TYPES}
\alias{FIBER_TYPES}
\title{Fiber-type-specific measures}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 672 rows and 11 columns.
}
\usage{
FIBER_TYPES
}
\description{
A \code{data.frame} containing fiber-type-specific measures:
  fiber type cross-sectional area (CSA) and fiber count.
}
\examples{
print.data.frame(head(FIBER_TYPES))

}
\seealso{
\code{\link{FIBER_TYPES_EMM}}, \code{\link{FIBER_TYPES_STATS}}
}
\keyword{datasets}
