% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{FIBER_TYPES_EMM}
\alias{FIBER_TYPES_EMM}
\title{Fiber-type-specific measures estimated marginal means}
\format{
An object of class \code{list} of length 1.
}
\usage{
FIBER_TYPES_EMM
}
\description{
A list of \code{emmGrid} objects: fiber type cross-sectional
  area (CSA).
}
\examples{
names(FIBER_TYPES_EMM) # available measures

str(FIBER_TYPES_EMM)

# Print one of the emmGrid objects
FIBER_TYPES_EMM[["Fiber CSA"]]

}
\seealso{
\code{\link{FIBER_TYPES}}, \code{\link{FIBER_TYPES_STATS}},
  \code{\link[emmeans]{emmeans}}
}
\keyword{datasets}
