% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{FIBER_TYPES_STATS}
\alias{FIBER_TYPES_STATS}
\title{Statistical analysis results of fiber-type-specific measures}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 56 rows and 20 columns.
}
\usage{
FIBER_TYPES_STATS
}
\description{
A \code{data.frame} containing statistical analysis results of
  fiber-type-specific measures: fiber type cross-sectional area (CSA).
}
\examples{
unique(FIBER_TYPES_STATS$response) # available measures

print.data.frame(head(FIBER_TYPES_STATS))

}
\seealso{
\code{\link{FIBER_TYPES}}, \code{\link{FIBER_TYPES_EMM}}
}
\keyword{datasets}
