% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MUSCLES}
\alias{MUSCLES}
\title{Muscle-specific measures}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1116 rows and 11 columns.
}
\usage{
MUSCLES
}
\description{
A \code{data.frame} containing muscle-type specific measures
  taken from medial gastrocnemius (MG), lateral gastroc (LG), plantaris (PL),
  and soleus (SOL): terminal muscle mass, mean cross-sectional area (CSA),
  glycogen, capillary contacts, and citrate synthase.
}
\examples{
print.data.frame(head(MUSCLES))

}
\seealso{
\code{\link{MUSCLES_EMM}}, \code{\link{MUSCLES_STATS}}
}
\keyword{datasets}
