% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MUSCLES_EMM}
\alias{MUSCLES_EMM}
\title{Muscle-specific measures estimated marginal means}
\format{
An object of class \code{list} of length 5.
}
\usage{
MUSCLES_EMM
}
\description{
A list of \code{emmGrid} objects, one for each muscle-specific
measures: terminal muscle mass, mean cross-sectional area (CSA), glycogen,
capillary contacts, and citrate synthase.
}
\examples{
names(MUSCLES_EMM) # available measures

str(MUSCLES_EMM)

# Print one of the emmGrid objects
MUSCLES_EMM[["Glycogen"]]

}
\seealso{
\code{\link{MUSCLES}}, \code{\link{MUSCLES_STATS}},
\code{\link[emmeans]{emmeans}}
}
\keyword{datasets}
