% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{PHYSIO}
\alias{PHYSIO}
\title{All physiology data}
\format{
An object of class \code{data.frame} with 282 rows and 123 columns.
}
\usage{
PHYSIO
}
\description{
All physiology measures combined into a single
  \code{data.frame}.
}
\keyword{datasets}
