% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_baseline.R
\name{plot_baseline}
\alias{plot_baseline}
\title{Plot baseline or pre-training measures}
\usage{
plot_baseline(
  x,
  response,
  conf,
  stats,
  sex = c("Female", "Male"),
  age = c("6M", "18M"),
  y_position = NULL
)
}
\arguments{
\item{x}{a \code{data.frame} with columns \code{"sex"}, \code{"age"},
\code{"group"}, and a column specified by \code{response}.}

\item{response}{character; the name of a column in \code{x} to plot.}

\item{conf}{a \code{data.frame} with confidence interval data.}

\item{stats}{optional \code{data.frame} with results of statistical analyses
filtered to the response of interest.}

\item{sex}{character; the sex of the samples to plot: "Female" or "Male".}

\item{age}{character; the age of the samples to plot: "6M" or "18M".}

\item{y_position}{numeric; manually specify starting y position of p-value
brackets.}
}
\description{
Plot baseline or pre-training measures.
}
