% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dissection_times.R
\name{plot_dissection_times}
\alias{plot_dissection_times}
\title{Plot tissue dissection times}
\usage{
plot_dissection_times(x, ymax = 12)
}
\arguments{
\item{x}{\code{data.frame} of dissection times.}

\item{ymax}{numeric; upper bound on y-axis (in minutes).}
}
\description{
Plot the tissue dissection times. This is the amount of time
elapsed between tissue collection and freezing or death and freezing (if
tissue collection was performed post-mortem).
}
