% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_lm.R
\name{plot_lm}
\alias{plot_lm}
\title{Regression diagnostic plots}
\usage{
plot_lm(model)
}
\arguments{
\item{model}{output of \code{lm} or \code{glm}.}
}
\description{
Regression diagnostic plots
}
\examples{
data("iris")
fit <- lm(Sepal.Length ~ ., data = iris)
plot_lm(fit)

}
