% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_lme.R
\name{plot_lme}
\alias{plot_lme}
\title{Linear mixed model diagnostic plots}
\usage{
plot_lme(model)
}
\arguments{
\item{model}{output of \code{\link[nlme]{lme}}.}
}
\description{
Linear mixed model diagnostic plots
}
