% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{PRE_POST_STATS}
\alias{PRE_POST_STATS}
\title{Statistical analysis results of post - pre training differences}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 90 rows and 20 columns.
}
\usage{
PRE_POST_STATS
}
\description{
A \code{data.frame} containing statistical analysis results of
  post - pre training differences: NMR body mass, lean mass, fat mass, %% lean
  mass, %% fat mass; difference between terminal and pre-training NMR mass at
  1W and 2W of endurance training; absolute VO\eqn{_2}max; relative
  VO\eqn{_2}max; and maximum run speed.
}
\examples{
unique(PRE_POST_STATS$response)

print.data.frame(head(PRE_POST_STATS))

}
\seealso{
\code{\link{PHYSIO}}
}
\keyword{datasets}
