% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstandard.mlm.R
\name{rstandard.mlm}
\alias{rstandard.mlm}
\title{rstandard method for "mlm" class}
\source{
\url{https://stackoverflow.com/a/39768104}
}
\usage{
\method{rstandard}{mlm}(model)
}
\arguments{
\item{model}{an \code{"mlm"} ("multivariate linear model") object. See
\code{\link[stats]{lm}} for details.}
}
\description{
rstandard method for "mlm" class
}
