% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{VO2MAX}
\alias{VO2MAX}
\title{VO2max test measures}
\format{
An object of class \code{data.frame} with 234 rows and 27 columns.
}
\usage{
VO2MAX
}
\description{
A \code{data.frame} containing measures from the pre- and
  post-training VO\eqn{_2}max tests: total body mass, absolute and relative
  VO\eqn{_2}max, respiratory exchange rate (RER), VCO\eqn{_2}, maximum run
  speed, blood lactate, and test duration.
}
\examples{
print.data.frame(head(VO2MAX))

}
\seealso{
\code{\link{PHYSIO}}
}
\keyword{datasets}
