% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{WEEKLY_BODY_MASS_EMM}
\alias{WEEKLY_BODY_MASS_EMM}
\title{Weekly body mass estimated marginal means}
\format{
An object of class \code{emmGrid} of length 1.
}
\usage{
WEEKLY_BODY_MASS_EMM
}
\description{
An \code{emmGrid} object for weekly body mass.
}
\examples{
str(WEEKLY_BODY_MASS_EMM)

}
\seealso{
\code{\link{WEEKLY_MEASURES}}, \code{\link{WEEKLY_BODY_MASS_STATS}},
\code{\link[emmeans]{emmeans}}
}
\keyword{datasets}
