% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{WEEKLY_BODY_MASS_STATS}
\alias{WEEKLY_BODY_MASS_STATS}
\title{Statistical analysis results of weekly body mass}
\format{
An object of class \code{data.frame} with 32 rows and 20 columns:

\describe{
\item{response}{character; the variable being tested.}
\item{age}{factor; the age of the rat at the beginning of the training
protocol. Levels: "6M" (Adult) and "18M" (Aged).}
\item{sex}{factor; the sex of the rat with levels "Female" and "Male".}
\item{week}{factor; the week the measurement was recorded relative to the
start of the training protocol (levels 1--8).}
\item{contrast}{factor; the comparison between groups. All contrasts are
ratios between the 8W and SED group means ("8W / SED").}
\item{estimate_type}{character; interpretation of the value in the
\code{estimate} column.}
\item{null}{numeric; the value of the estimate under the null hypothesis.}
\item{estimate}{numeric; }
\item{SE}{numeric; the standard error of the ratio estimate.}
\item{lower.CL}{numeric; lower bound of the 95\% confidence interval.}
\item{upper.CL}{numeric; upper bound of the 95\% confidence interval.}
\item{statistic_type}{character; the type of statistical test.}
\item{statistic}{numeric; the value of the test statistic.}
\item{df}{numeric; the number of residual degrees of freedom.}
\item{p.value}{numeric; the p-value associated with the statistical
test.}
\item{p.adj}{numeric; the Holm-adjusted p-value.}
\item{signif}{logical; \code{TRUE} if \code{p.adj} < 0.05.}
\item{model_type}{character; the statistical model used for testing.}
\item{formula}{character; the model formula. Includes the response
variable, any transformations, and predictors.}
\item{correlation}{character; the within-group correlation structure. All
"corAR1(form = ~1 | pid)".}
}
}
\usage{
WEEKLY_BODY_MASS_STATS
}
\description{
Statistical analysis results of weekly body mass.
}
\examples{
head(WEEKLY_BODY_MASS_STATS)

}
\seealso{
\code{\link{WEEKLY_MEASURES}}, \code{\link{WEEKLY_BODY_MASS_EMM}}
}
\keyword{datasets}
