% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{WEEKLY_MEASURES}
\alias{WEEKLY_MEASURES}
\title{Weekly body weight and blood lactate measures}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 2226 rows and 11 columns.
}
\usage{
WEEKLY_MEASURES
}
\description{
A \code{data.frame} containing weekly measures of total body
  mass and blood lactate.
}
\examples{
print.data.frame(head(WEEKLY_MEASURES))

}
\seealso{
\code{\link{PHYSIO}}
}
\keyword{datasets}
