% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{WEIGHTS}
\alias{WEIGHTS}
\title{Body mass recorded on different days}
\format{
An object of class \code{data.frame} with 282 rows and 12 columns.
}
\usage{
WEIGHTS
}
\description{
A \code{data.frame} containing measures of total body mass (in
  grams) recorded on different days (pre- and post-training measures
  included, if applicable): NMR, VO\eqn{_2}max, and day of sacrifice.
}
\examples{
print.data.frame(head(WEIGHTS))

}
\seealso{
\code{\link{PHYSIO}}
}
\keyword{datasets}
