% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ANALYTES}
\alias{ANALYTES}
\title{Plasma clinical analytes}
\format{
An object of class \code{data.frame} with 282 rows and 20 columns.
}
\usage{
ANALYTES
}
\description{
A \code{data.frame} containing common clinical analytes measured
  in plasma.
}
\examples{
print.data.frame(head(ANALYTES))

}
\seealso{
\code{\link{ANALYTES_EMM}}, \code{\link{ANALYTES_STATS}}
}
\keyword{datasets}
