% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ANALYTES_EMM}
\alias{ANALYTES_EMM}
\title{Estimated marginal means object for each plasma clinical analyte}
\format{
An object of class \code{list} of length 9.
}
\usage{
ANALYTES_EMM
}
\description{
A list of \code{emmGrid} objects, one for each clinical analyte.
}
\keyword{datasets}
