% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ANALYTES_STATS}
\alias{ANALYTES_STATS}
\title{Results of statistical analysis of plasma clinical analytes}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 136 rows and 18 columns.
}
\usage{
ANALYTES_STATS
}
\description{
A \code{data.frame} containing analysis results for common
  clinical analytes measured in plasma.
}
\keyword{datasets}
