% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{BASELINE_STATS}
\alias{BASELINE_STATS}
\title{Statistical analysis results of baseline (pre-training) measures}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 112 rows and 24 columns.
}
\usage{
BASELINE_STATS
}
\description{
A \code{data.frame} containing statistical analysis results of
  baseline (pre-training) measures: NMR mass, lean mass, fat mass, % lean
  mass, % fat mass; absolute VO\eqn{_2}max; relative VO\eqn{_2}max; and
  maximum run speed.
}
\examples{
unique(BASELINE_STATS$response) # available measures

print.data.frame(head(BASELINE_STATS))

}
\seealso{
\code{\link{BASELINE_EMM}}
}
\keyword{datasets}
