% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_lm.R
\name{plot_lm}
\alias{plot_lm}
\title{Show all regression diagnostic plots}
\usage{
plot_lm(model)
}
\arguments{
\item{model}{output of \code{lm} or \code{glm}.}
}
\description{
Show all regression diagnostic plots
}
